requirements.txt
pytorch-lightning==1.4.0
torch==1.9.0
hydra-core==1.1.0
hydra-colorlog==1.1.0
hydra-joblib-launcher==1.1.5
hydra-optuna-sweeper==1.1.0
mlflow==1.19.0
pandas==1.3.1
numpy==1.19.5
scikit-learn==0.24.2
matplotlib==3.4.2
seaborn==0.11.1